refactor: refactored errors

+++ internal/core/servises/code_example.go
@@ -4,8 +4,8 @@ import (
 	"code-typing-text-service/internal/core/domain"
 	"code-typing-text-service/internal/core/ports"
 	"code-typing-text-service/internal/core/ports/dto"
-	"code-typing-text-service/internal/core/ports/errors"
 	"code-typing-text-service/pkg/logging"
+	"fmt"
 	"github.com/jinzhu/copier"
 )
 
@@ -26,27 +26,23 @@ func (s *CodeExampleService) GetProgrammingLanguages() (getProgrammingLanguagesD
 
 	err = copier.Copy(&getProgrammingLanguagesDto, &programmingLanguages)
 	if err != nil {
-		err = &errors.MappingError{Message: `struct mapping error`}
+		err = fmt.Errorf(`struct mapping error: %w`, ports.InternalServerError)
 	}
 	return
 }
 
-func (s *CodeExampleService) GetCodeExampleByUUID(userID, UUID string) (getCodeExampleDto dto.GetCodeExampleDto, err error) {
-	var codeExample domain.CodeExample
-	codeExample, err = s.repo.GetCodeExampleByUUID(UUID)
-	if err != nil {
-		return
-	}
-	if codeExample.UserID != userID && codeExample.UserID != "" {
-		err = &errors.NotFoundError{Message: `code example not found`}
-		return
+func (s *CodeExampleService) GetCodeExampleByUUID(userID, UUID string) (dto.GetCodeExampleDto, error) {
+	codeExample, err := s.repo.GetCodeExampleByUUID(UUID)
+	if err != nil || codeExample.UserID != userID && codeExample.UserID != "" {
+		return dto.GetCodeExampleDto{}, fmt.Errorf(`code example not found: %w`, ports.NotFoundError)
 	}
 
+	var getCodeExampleDto dto.GetCodeExampleDto
 	err = copier.Copy(&getCodeExampleDto, &codeExample)
 	if err != nil {
-		err = &errors.MappingError{Message: `struct mapping error`}
+		err = fmt.Errorf(`struct mapping error: %w`, ports.InternalServerError)
 	}
-	return
+	return getCodeExampleDto, nil
 }
 
 func (s *CodeExampleService) GetCodeExamples(userID string) (getCodeExamplesDto []dto.GetCodeExampleDto, err error) {
@@ -54,27 +50,29 @@ func (s *CodeExampleService) GetCodeExamples(userID string) (getCodeExamplesDto 
 
 	err = copier.Copy(&getCodeExamplesDto, &codeExamples)
 	if err != nil {
-		err = &errors.MappingError{Message: `struct mapping error`}
+		err = fmt.Errorf(`struct mapping error: %w`, ports.InternalServerError)
 	}
 	return
 }
 
-func (s *CodeExampleService) GetCodeExamplesByProgrammingLanguageName(userID, programmingLanguageName string) (getCodeExamplesDto []dto.GetCodeExampleDto, err error) {
+func (s *CodeExampleService) GetCodeExamplesByProgrammingLanguageName(userID, programmingLanguageName string) ([]dto.GetCodeExampleDto, error) {
 	codeExamples, err := s.repo.GetCodeExamplesByProgrammingLanguageName(userID, programmingLanguageName)
 	if err != nil {
 		s.log.Infof(`error getting code examples by programming language '%s' due to error: %v`, programmingLanguageName, err)
-		return
+		err = fmt.Errorf(`code examples not found: %w`, ports.NotFoundError)
+		return nil, err
 	}
 
+	var getCodeExamplesDto []dto.GetCodeExampleDto
 	err = copier.Copy(&getCodeExamplesDto, &codeExamples)
 	if err != nil {
-		err = &errors.MappingError{Message: `struct mapping error`}
+		return nil, fmt.Errorf(`struct mapping error: %w`, ports.InternalServerError)
 	}
-	return
+	return getCodeExamplesDto, nil
 }
 
 func (s *CodeExampleService) CreateCodeExample(userID string, createCodeExampleDto dto.CreateCodeExampleDto) (string, error) {
-	UUID, err := s.repo.SaveCodeExample(
+	UUID, err := s.repo.CreateCodeExample(
 		domain.CodeExample{
 			UserID:                  userID,
 			ProgrammingLanguageUUID: createCodeExampleDto.ProgrammingLanguageUUID,
@@ -83,24 +81,17 @@ func (s *CodeExampleService) CreateCodeExample(userID string, createCodeExampleD
 	)
 	if err != nil {
 		s.log.Infof(`error creating code example due to error: %v`, err)
-		return "", err
+		return "", fmt.Errorf(`error creating code example: %w`, ports.BadRequestError)
 	}
 
 	return UUID, nil
 }
 
-func (s *CodeExampleService) DeleteCodeExample(userID, UUID string) (err error) {
+func (s *CodeExampleService) DeleteCodeExample(userID, UUID string) error {
 	codeExample, err := s.repo.GetCodeExampleByUUID(UUID)
-	if err != nil {
-		return
+	if err != nil || codeExample.UserID != userID {
+		return fmt.Errorf(`code example not found: %w`, ports.NotFoundError)
 	}
-	if codeExample.UserID != userID {
-		err = &errors.NoAccessError{Message: `no access to code example`}
-		return
-	}
-	err = s.repo.DeleteCodeExample(UUID)
-	if err != nil {
-		s.log.Infof(`error deleting code example due to error: %v`, err)
-	}
-	return err
+	s.repo.DeleteCodeExample(UUID)
+	return nil
 }
